+++ src/main/java/com/phion/spring/ioc/context/ApplicationContext.java
@@ -1,0 +1,11 @@
+package com.phion.spring.ioc.context;
+
+import com.phion.spring.ioc.beans.factory.BeanFactory;
+
+/**
+ * 应用上下文，spring功能集结地
+ *
+ * @author yanful
+ */
+public interface ApplicationContext extends BeanFactory {
+}